feat(stella-tools): make repo_status and project_overview answer the questions callers actually have
repo_status reported branch/ahead/behind/changed -- all WORKING-TREE state.
A caller asking about lost work got a truthful, clean, useless answer: on
Terminal-Bench fix-git it returned `branch: null` (detached) and `changed: []`
(the snapshot seal had just committed everything), which is correct and says
nothing, so the agent went to raw `git fsck`/`git reflog` from step 1. No field
in the payload could ever have expressed "a commit exists and nothing points at
it".

Adds:

- `head` -- where HEAD points, named. The only way to read the state at all
  when `branch` is null.
- `unreachable` -- commits the reflog reaches that no branch, tag or remote
  does: work orphaned by a checkout, reset or abandoned rebase. A ref walk
  (`rev-list --all --reflog --not --branches --tags --remotes`) rather than
  `fsck --unreachable`, which scans every object in the store; only one of
  those is affordable on every call.
- `stashes` -- zero is a fact worth having. It is how a caller rules the stash
  out rather than guessing.

The pipeline's own baseline snapshots are excluded by author email. They are
unreachable by construction -- a candidate workspace is a detached-HEAD shadow
-- so listing them would present Stella's bookkeeping to the agent as the
user's lost work, which is exactly what made an authored witness unsatisfiable
on the same task (PR #2537). Filtered in Rust: git can filter TO an author but
has no invert-author, and `--invert-grep` applies only to `--grep`.

project_overview gains:

- `manifests` -- every package-manager file the script index parsed, with its
  path. `scripts` reports the verbs those manifests bind, which answers "how do
  I run this"; it cannot answer "what kind of project is this and where are its
  package boundaries" -- a monorepo rendered as two runner names.
- `git` -- remotes, branch count, and the naming convention the existing
  branches imply. A new branch should look like the ones beside it, and the
  repository is the only place that answer lives when no context record states
  one. Reported with the share of branches following it, because a convention
  two of nine branches use is a coincidence; nothing dominant reports null
  rather than guessing, since a caller told "the convention is x/" will follow
  it.

repo.rs sat 8 lines under the 1500-line limit, so this crossed it. Tests move
to a sibling repo/tests.rs -- a NEW god file fails the gate outright rather
than being grandfathered, and no baseline entry was added.

Refs #2374
